Create fork.c
child and parent can be distinguished by the value returned by fork

If fork() returns a negative value, the creation of a child process was unsuccessful.
fork() returns a zero to the newly created child process.
fork() returns a positive value, the process ID of the child process, to the parent. The returned process ID is of type pid_t defined in sys/types.h. 
Normally, the process ID is an integer. Moreover, a process can use function getpid() to retrieve the process ID assigned to this process,
getppid() used to get pid of parent process.